Move minor dictionary helpers to header files to solve linkage problems and delete impl-files thus superfluous.
