+++ sli/lockptrdatum.h
@@ -117,7 +117,45 @@ private:
   bool operator==( lockPTR< D >& );
 };
 
+/* equals(const Datum* datum)
+   returns: true if *this and Datum *dat are both lockptr references
+   to the same underlying object.
 
-/******************************************/
+   The definition of the equals method assumes that no further
+   distinguishing data is added by derivation.  Aka, the template
+   class is never inherited from, and therefore type equality is
+   guaranteed by template parameter equality.
+*/
+template < class D, SLIType* slt >
+bool
+lockPTRDatum< D, slt >::equals( const Datum* dat ) const
+{
+  const lockPTRDatum< D, slt >* ddc = dynamic_cast< const lockPTRDatum< D, slt >* >( dat );
+  return ddc && lockPTR< D >::operator==( *ddc );
+}
+
+template < class D, SLIType* slt >
+void
+lockPTRDatum< D, slt >::pprint( std::ostream& out ) const
+{
+  out << "<lockPTR[" << this->references() << "]->" << this->gettypename() << '(' << static_cast< void* >( this->get() )
+      << ")>";
+  this->unlock();
+}
+
+template < class D, SLIType* slt >
+void
+lockPTRDatum< D, slt >::print( std::ostream& out ) const
+{
+  out << '<' << this->gettypename() << '>';
+}
+
+template < class D, SLIType* slt >
+void
+lockPTRDatum< D, slt >::info( std::ostream& out ) const
+{
+  //  out << *dynamic_cast<C *>(const_cast<lockPTR<C,slt> *>(this));
+  pprint( out );
+}
 
 #endif
