+++ sli/numericdatum.h
@@ -33,7 +33,7 @@
 #include "genericdatum.h"
 
 
-// prefixed all references to members of GenericDatum with this->,
+// Prefixed all references to members of GenericDatum with this->,
 // since HP's aCC otherwise complains about them not being declared
 // according to ISO Standard Sec. 14.6.2(3) [temp.dep]
 // HEP, 2001-08-08
@@ -75,9 +75,17 @@ public:
     return d;
   }
 
-  void input_form( std::ostream& ) const;
-  void pprint( std::ostream& ) const;
+  void
+  input_form( std::ostream& o ) const
+  {
+    pprint( o );
+  }
 
+  void
+  pprint( std::ostream& o ) const
+  {
+    this->print( o );
+  }
 
   static void* operator new( size_t size )
   {
