+++ sli/symboldatum.cc
@@ -22,12 +22,6 @@
 
 #include "symboldatum.h"
 
-// explicit template instantiation needed
-// because otherwise methods defined in
-// numericdatum_impl.h will not be instantiated
-// Moritz, 2007-04-16
-template class AggregateDatum< Name, &SLIInterpreter::Symboltype >;
-
 
 // initialization of static members requires template<>
 // see Stroustrup C.13.1 --- HEP 2001-08-09
